refactor static servers to allow altnernative implementations
