+++ conductor_api/src/conductor/base.rs
@@ -53,7 +53,8 @@ use holochain_net::{
 };
 use interface::{ConductorApiBuilder, InstanceMap, Interface};
 use signal_wrapper::SignalWrapper;
-use static_file_server::StaticServer;
+use static_file_server::{ConductorStaticFileServer};
+use static_server_impls::HyperStaticServer as StaticServer;
 
 lazy_static! {
     /// This is a global and mutable Conductor singleton.
